backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
supabase>=2.15
pydantic>=2.10
pydantic-settings>=2.7
PyJWT[crypto]>=2.8
httpx>=0.28
